Engine: Remove old processing graph implementation
