--- engine/src/modules/processing_graph/compiler/graph_compilation_result.h
@@ -1,74 +1,0 @@
-/*
-  Copyright (C) 2024 - 2026 Joshua Wade
-
-  This file is part of Anthem.
-
-  Anthem is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  Anthem is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with Anthem. If not, see <https://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#include "modules/processing_graph/compiler/graph_action.h"
-#include "modules/processing_graph/compiler/graph_process_context.h"
-#include "modules/processing_graph/model/node.h"
-
-#include <juce_core/juce_core.h>
-#include <memory>
-#include <vector>
-
-// This class is used to represent the result of compiling a processing graph.
-namespace anthem {
-
-class GraphCompilationResult {
-private:
-  JUCE_LEAK_DETECTOR(GraphCompilationResult)
-public:
-  // Actions in execution order for the compiled graph.
-  std::vector<GraphAction> actions;
-
-  // Shared execution data for the compiled graph.
-  float sampleRate = 0.0f;
-
-  // Owns all graph-scoped runtime storage for this compiled graph, including
-  // the node process contexts that point into that storage.
-  std::unique_ptr<GraphProcessContext> graphProcessContext;
-
-  // This contains a shared_ptr reference to each graph node that was present
-  // when this context was created.
-  //
-  // This exists to ensure that graph nodes are not freed until the audio thread
-  // is done with them. The audio thread itself can't interact with these since
-  // they aren't thread-safe; instead, the deallocation of the compilation
-  // result on the main thread triggers deallocation for any nodes that need it.
-  //
-  // Note that the audio thread only has access to these nodes so that it can
-  // call node->processor->process(&context, numSamples).
-  //
-  // If this list didn't exist, then we'd risk the following use-after-free:
-  //    1. A node is created and connected to the rest of the graph.
-  //    2. The graph is compiled, and the result is sent to the audio thread.
-  //    3. The node is removed from the audio graph, with a compiled update
-  //       expected to be sent next.
-  //    4. Before the compiled update can be sent, this node is deallocated
-  //       because it has no more shared_ptr references.
-  //    5. The audio thread continues to use its raw pointer to try to access
-  //       the node, which results in a use-after-free.
-  std::vector<std::shared_ptr<Node>> graphNodes;
-
-  // Clean up the compilation result. This must be called before the compilation
-  // result is deallocated.
-  void cleanup();
-};
-
-} // namespace anthem
