+++ engine/src/modules/processing_graph_threaded/executor/graph_executor_shared.cpp
@@ -19,10 +19,10 @@
 
 #include "graph_executor_shared.h"
 
-#include "modules/processing_graph/compiler/node_process_context.h"
 #include "modules/processing_graph/processor/event_buffer.h"
 #include "modules/processing_graph/processor/processor.h"
 #include "modules/processing_graph_threaded/model/runtime_graph.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 #include <atomic>
 #include <juce_core/juce_core.h>
