+++ engine/src/modules/processing_graph_threaded/graph_processor.cpp
@@ -19,8 +19,8 @@
 
 #include "graph_processor.h"
 
-#include "modules/processing_graph/runtime/graph_runtime_services.h"
 #include "modules/processing_graph_threaded/executor/graph_executor.h"
+#include "modules/processing_graph_threaded/runtime/graph_runtime_services.h"
 #include "modules/util/intentionally_leak.h"
 
 namespace anthem::threaded_graph {
