+++ engine/src/modules/processing_graph_threaded/runtime/live_note_id_generator.h
