+++ engine/src/modules/processors/balance.cpp
@@ -19,7 +19,7 @@
 
 #include "balance.h"
 
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 #include <juce_core/juce_core.h>
