+++ engine/src/modules/processors/db_meter.cpp
@@ -21,7 +21,7 @@
 
 #include "modules/core/engine.h"
 #include "modules/core/visualization/visualization_broker.h"
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 #include <algorithm>
