+++ engine/src/modules/processors/gain.cpp
@@ -19,7 +19,7 @@
 
 #include "gain.h"
 
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 namespace anthem {
