+++ engine/src/modules/processors/live_event_provider.cpp
@@ -19,7 +19,7 @@
 
 #include "live_event_provider.h"
 
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 namespace anthem {
