+++ engine/src/modules/processors/master_output.cpp
@@ -20,7 +20,7 @@
 #include "master_output.h"
 
 #include "modules/core/engine.h"
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 #include <iostream>
 #include <juce_audio_basics/juce_audio_basics.h>
