+++ engine/src/modules/processors/sequence_note_provider.cpp
@@ -20,7 +20,7 @@
 #include "sequence_note_provider.h"
 
 #include "modules/core/engine.h"
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 #include "modules/sequencer/runtime/runtime_sequence_store.h"
 
 namespace anthem {
