+++ engine/src/modules/processors/simple_midi_generator.cpp
@@ -20,7 +20,7 @@
 #include "simple_midi_generator.h"
 
 #include "modules/core/engine.h"
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 #include "modules/sequencer/events/event.h"
 
 namespace anthem {
