+++ engine/src/modules/processors/vst3_processor.cpp
@@ -23,7 +23,7 @@
 
 #include "generated/lib/model/model.h"
 #include "modules/core/engine.h"
-#include "modules/processing_graph/compiler/node_process_context.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context.h"
 
 #if JUCE_WINDOWS
 // JUCE exposes the helper for matching the current thread's DPI-awareness
