+++ engine/test/modules/processing_graph/model/processing_graph_model_helpers_test.h
@@ -20,10 +20,10 @@
 #pragma once
 
 #include "generated/lib/engine_api/messages/messages.h"
-#include "modules/processing_graph/compiler/graph_process_context.h"
 #include "modules/processing_graph/graph_test_helpers.h"
 #include "modules/processing_graph/model/node.h"
-#include "modules/processing_graph/runtime/graph_runtime_services.h"
+#include "modules/processing_graph_threaded/runtime/graph_process_context.h"
+#include "modules/processing_graph_threaded/runtime/graph_runtime_services.h"
 #include "modules/processors/gain.h"
 
 #include <juce_core/juce_core.h>
