+++ engine/test/modules/processing_graph_threaded/runtime/node_process_context_test.h
@@ -20,9 +20,9 @@
 #pragma once
 
 #include "modules/core/constants.h"
-#include "modules/processing_graph/compiler/graph_process_context.h"
 #include "modules/processing_graph/graph_test_helpers.h"
-#include "modules/processing_graph/runtime/graph_runtime_services.h"
+#include "modules/processing_graph_threaded/runtime/graph_process_context.h"
+#include "modules/processing_graph_threaded/runtime/graph_runtime_services.h"
 
 #include <juce_core/juce_core.h>
