+++ engine/test/modules/processors/live_event_provider_test.h
@@ -19,9 +19,9 @@
 
 #pragma once
 
-#include "modules/processing_graph/compiler/graph_process_context.h"
 #include "modules/processing_graph/graph_test_helpers.h"
-#include "modules/processing_graph/runtime/graph_runtime_services.h"
+#include "modules/processing_graph_threaded/runtime/graph_process_context.h"
+#include "modules/processing_graph_threaded/runtime/graph_runtime_services.h"
 #include "modules/processors/live_event_provider.h"
 
 #include <juce_core/juce_core.h>
