+++ engine/test/test.cpp
@@ -19,15 +19,11 @@
 
 #include "console_logger.h"
 #include "modules/core/sequencer_test.h"
-#include "modules/processing_graph/compiler/graph_compiler_node_test.h"
-#include "modules/processing_graph/compiler/graph_compiler_test.h"
-#include "modules/processing_graph/compiler/graph_process_context_test.h"
-#include "modules/processing_graph/compiler/node_process_context_test.h"
 #include "modules/processing_graph/model/processing_graph_model_helpers_test.h"
 #include "modules/processing_graph/processor/event_buffer_test.h"
-#include "modules/processing_graph/runtime/graph_action_executor_test.h"
-#include "modules/processing_graph/runtime/graph_processor_test.h"
 #include "modules/processing_graph_threaded/model/runtime_graph_test.h"
+#include "modules/processing_graph_threaded/runtime/graph_process_context_test.h"
+#include "modules/processing_graph_threaded/runtime/node_process_context_test.h"
 #include "modules/processors/balance_test.h"
 #include "modules/processors/db_meter_test.h"
 #include "modules/processors/gain_parameter_mapping_test.h"
